+++ CHANGELOG.md
@@ -40,6 +40,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Improvements
 
 * Remove the warnings about some config settings [2095](https://github.com/provenance-io/provenance/pull/2095).
+* Record several scope NAVs with the umber upgrade [#2085](https://github.com/provenance-io/provenance/pull/2085).
+* Store the Figure Funding Trading Bridge Smart Contract as part of the umber upgrade [2102](https://github.com/provenance-io/provenance/pull/2102).
+
+### Dependencies
+
+- Bump `github.com/cometbft/cometbft` from v0.38.10-pio-1 to v0.38.10 ([#2097](https://github.com/provenance-io/provenance/pull/2097))
 
 ---